use AUR package ttf-font-awesome-4 for glyphs
the 'tt-font-awesome' package in the community repository is Font
Awesome v5, which is the version that the developers decided to split
their fonts into a free and nonfree group. The glyphs offered for free
in version 5 are solid, and not a complete match with those that were
offered in version 4.