restructure variable files
the intent is to keep generic configuration in as 'group variables', and
options which may potentially be host-specific in the configuration file
for 'host variables', like the hostname. the host variable file should
probably eventually be removed from the repository, and replaced with a
'*.example' file that users must copy and modify for their particular
environment.