add hostname role
i'm unhappy with the limited testing here, but due to ansible's atomic
nature, `/etc/hosts` cannot be modified within a docker container.

see https://github.com/ansible/ansible/issues/13981

we could use the `unsafe_writes` property to get around this, but we
definitely do not want that option enabled for live use. since options
for task actions can't be explicitly added via the command line or based
on a conditional, we'd be stuck with maintaining two copies of the tasks
that use `file` modules... and nobody wants that.