feat(timezone): support automatically updating /etc/localtime
This commit adds support for querying a remote endpoint to determine the
system timezone based on the current IP address. This, in conjunction
with the usage of any NTP solution (such as systemd-timesyncd), enables
automated updates to the system timezone and time. This is especially
useful on mobile devices, such as laptops.

closes #10